+++ pkg/analysis_server/test/analysis/notification_analysis_options_test.dart
@@ -28,7 +28,7 @@ class AnalysisOptionsFileNotificationTest extends PubPackageAnalysisServerTest {
   final testSource = '''
 void f() {
   var x = '';
-  int y = x; // Not assignable in strong-mode
+  int y = x; // Not assignable.
   print(y);
 }''';