+++ pkg/analysis_server/test/analysis_server_base.dart
@@ -27,8 +27,6 @@ import 'src/utilities/mock_packages.dart';
 /// TODO(scheglov) this is duplicate
 class AnalysisOptionsFileConfig {
   final List<String> experiments;
-  final bool implicitCasts;
-  final bool implicitDynamic;
   final List<String> lints;
   final bool strictCasts;
   final bool strictInference;
@@ -36,8 +34,6 @@ class AnalysisOptionsFileConfig {
 
   AnalysisOptionsFileConfig({
     this.experiments = const [],
-    this.implicitCasts = true,
-    this.implicitDynamic = true,
     this.lints = const [],
     this.strictCasts = false,
     this.strictInference = false,
@@ -58,9 +54,6 @@ class AnalysisOptionsFileConfig {
     buffer.writeln('    strict-casts: $strictCasts');
     buffer.writeln('    strict-inference: $strictInference');
     buffer.writeln('    strict-raw-types: $strictRawTypes');
-    buffer.writeln('  strong-mode:');
-    buffer.writeln('    implicit-casts: $implicitCasts');
-    buffer.writeln('    implicit-dynamic: $implicitDynamic');
 
     buffer.writeln('linter:');
     buffer.writeln('  rules:');