+++ pkg/analyzer/test/src/diagnostics/field_initializer_not_assignable_test.dart
@@ -10,8 +10,6 @@ import '../dart/resolution/context_collection_resolution.dart';
 main() {
   defineReflectiveSuite(() {
     defineReflectiveTests(FieldInitializerNotAssignableTest);
-    defineReflectiveTests(
-        FieldInitializerNotAssignableWithoutNullSafetyAndNoImplicitCastsTest);
     defineReflectiveTests(FieldInitializerNotAssignableWithStrictCastsTest);
   });
 }
@@ -64,22 +62,6 @@ enum E {
       error(CompileTimeErrorCode.CONST_CONSTRUCTOR_FIELD_TYPE_MISMATCH, 11, 1),
       error(CompileTimeErrorCode.FIELD_INITIALIZER_NOT_ASSIGNABLE, 47, 2),
       error(CompileTimeErrorCode.CONST_FIELD_INITIALIZER_NOT_ASSIGNABLE, 47, 2),
-    ]);
-  }
-}
-
-@reflectiveTest
-class FieldInitializerNotAssignableWithoutNullSafetyAndNoImplicitCastsTest
-    extends PubPackageResolutionTest
-    with WithoutNullSafetyMixin, WithNoImplicitCastsMixin {
-  test_constructorInitializer() async {
-    await assertErrorsWithNoImplicitCasts('''
-class A {
-  int i;
-  A(num n) : i = n;
-}
-''', [
-      error(CompileTimeErrorCode.FIELD_INITIALIZER_NOT_ASSIGNABLE, 36, 1),
     ]);
   }
 }